backend/requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
rest-framework-simplejwt==0.0.2
simplejwt==2.0.1
sqlparse==0.4.4
typing==3.7.4.3
